Dimensions to int64_t and begin_step

+++ bindings/fortran/f2c/adios2_f2c_variable.cpp
@@ -61,7 +61,7 @@ void FC_GLOBAL(adios2_variable_type_f2c,
 
 void FC_GLOBAL(adios2_set_shape_f2c,
                ADIOS2_SET_SHAPE_F2C)(adios2_Variable **variable,
-                                     const int *ndims, const int *shape,
+                                     const int *ndims, const int64_t *shape,
                                      int *ierr)
 {
     *ierr = 0;
@@ -84,10 +84,10 @@ void FC_GLOBAL(adios2_set_shape_f2c,
 
 void FC_GLOBAL(adios2_set_selection_f2c,
                ADIOS2_SET_SELECTION_F2C)(adios2_Variable **variable,
-                                         const int *ndims, const int *start,
-                                         const int *count, int *ierr)
+                                         const int *ndims, const int64_t *start,
+                                         const int64_t *count, int *ierr)
 {
-    auto lf_IntToSizeT = [](const int *dimensions, const int size,
+    auto lf_IntToSizeT = [](const int64_t *dimensions, const int size,
                             std::vector<std::size_t> &output) {
 
         output.resize(size);
